設定リポジトリ中に lucie_config.rb がある場合、config/lucie_config.rb_example よりも優先するように修正。

+++ app/models/installers.rb
@@ -70,9 +70,13 @@ class Installers
 
   def write_config_example installer
     lucie_config_example = File.join( RAILS_ROOT, 'config', 'lucie_config.rb_example' )
+    lucie_config_in_subversion = File.join( installer.path, 'work', 'lucie_config.rb' )
     lucie_config = File.join( installer.path, 'lucie_config.rb' )
-    FileUtils.cp lucie_config_example, lucie_config
+    if File.exists?( lucie_config_example ) and not File.exists?( lucie_config_in_subversion )
+      FileUtils.cp lucie_config_example, lucie_config
+    end
 
+    # XXX 
     rakefile_example = File.join( RAILS_ROOT, 'config', 'Rakefile_example' )
     rakefile = File.join( installer.path, 'Rakefile' )
     FileUtils.cp rakefile_example, rakefile
